bittorrent/main.py: Make piece scrambling conditional
The previous patch implemented piece scrambling to avoid sending
always an infinite stream of 'A' chars to the other side.  The next
patch wants to obfuscate the whole protocol to perform a test that
looks like the one of Glasnost.

But it seems a waste to scramble the pieces and then obfuscate the
protocol and it would certainly be better to avoid the scrambling when
we obfuscate.

So, this patch makes scrambling conditional, and we will turn it
off when we obfuscate the whole protocol.