bittorrent/main.py: Allow to obfuscate the whole protocol
This patch allows to toggle a switch that causes the whole protocol
to be obfuscated.  This is similar to what the original BitTorrent code
did, but we don't exchange a couple of keys using D-H here.  Instead,
we assume that the encryption keys will be decided offline, i.e.
during the negotiate phase.

We might want to exploit this feature to implement a test similar
to one of the tests provided by Glasnost, that allows to compare the
speed of the bare protocol with the one of the obfuscated protocol.