bittorrent/main.py: Keep at most 32 pieces in flight
This commit modifies the downloader and now the code attempts to
keep at most 32 pieces in flight, i.e. one MByte.  This number is
kind of arbitrary and might indeed be wrong because it can limit
the total throughput that one can achieve.  Nonetheless I will commit
this patch to be able to do some more serious testing.