setup.py: Rewritten to drive the win32 build process
This patch rewrites setup.py to be able to drive the whole win32
build process, so we can get rid of the (ugly) batch files.

This patch is the first one in a branch that aims to simplify the
win32 build process and to build, at the same time, a console and a
windows neubot executables.  More in detail, this branch contains
the following patches:

1. setup.py: Rewritten to drive the win32 build process
2. Make-*.bat: Remove the now-unneeded batch files
3. Revert "log.py: With windows force background logger"
4. bin/neubotw: Script that becomes the windows executable
5. bin/neubot: Import instead of from..import for stdlib
6. setup.py: Build both console and windows executables
7. neubot.nsi: Use neubotw.exe instead of neubot.exe

The rewrite of setup.py that is proposed in this patch comes after
another round of study of the docutils documentation.  In rewriting
the file, I believe I have also built a foundation that will allow
to install and package Neubot for Unix using setup.py too.  This isn't
possible as of now, because some details must be sorted out, but I
envision the feasibility of that -- and one obvious advantage is
that it will allow to create neubot RPMs.

Please note that the batch files used to invoke py2exe in a way
that embeds most of the pythonic stuff into <library.zip>.  The new
code does not do that and therefore the Win32 distribution dir
will contain more files.