Revert "log.py: With windows force background logger"
This reverts commit 579913f59ffdcb1d89e93845b3c225f7ae3be876.

I have found a way to compile both the console and the windows
executable at the same time, but cannot amend the build process to
do that, since there is an hack that forces Neubot to always log
on the NT EventLog.  So, this patch undoes the hack, for the console
executable to be able to log to the standard error.