bin/neubotw: Script that becomes the windows executable
This script is same as bin/neubot except that it immediately re-
directs the logger to the NT EventLog Handler.  This is the script
that will be used to make the windows executable, i.e. the one
that does not have an attached console.

The decision is to name neubotw the windows version and neubot the
console version, for simmetry with pythonw.exe/python.exe (I was not
aware of that 'convention' but now that I've noticed it, I admit it
makes sense -- and I like it).