bin/neubot: Import instead of from..import for stdlib
Use import foo instead of from foo import bar when foo is part of
the standard library to agument code readability.  So a new programmer
knows that when there is not a namespace it must be something that is
part of Neubot.  Otherwise, it's standard library.

After this commit the diff between bin/neubot and bin/neubotw should
consist of just a couple of lines.