www/results.html: unbreak plots for internet explorer
Somehow we forgot to update the line that fetches the canvas for
internet explorer, and so it was not possible to plot the speed
with such family of browsers.

This commit fixes _also_ index.html that contains the same error,
even if that was not visible, because there are not plots in index.html
at the moment.