js/results.js: Break x axis lables to unbreak zoom
This patch temporarily breaks x axis labels to unbreak the zoom,
which does not work very well with current labels.

This is the first patch in a branch that aims to adjust -- in a
quick and dirty way -- the results plot just before we release
the next version of Neubot.  The quick and dirty commitment stems
from the fact that Alessio Palmero is already at work to refine
to whole web interface, but his changes are scheduled for merging
after the release.  Thus, I would like that merge not to become
a blood bath.  More in detail, this branch contains the following
list of patches:

1. js/results.js: Break x axis labels to unbreak zoom
2. js/ += jqplot.dateAxisRenderer.min.js
3. www/results.html: Import from jqplot.dateAxisRenderer
4. js/results.js: Quick-n-dirty hack to prettify the x axis labels
5. js/results.js: Remove old-n-commented-out x-axis-labels code