js/results.js: Quick-n-dirty hack to prettify the x axis labels
This patch is very dirty but fixes the problem of printing x axis
labels vs. the broken zoom.  I certainly could have written something
less dirty but there is a reason not to do that now.  Alessio is working
on the web user interface, to simplify the code and his contribution is
to be merged _after_ the next release.  So I don't want to make too many
changes now for the merge to be as smooth as possible.