neubot/speedtest.py: First calibrate, then measure
This patch rewrites the speedtest algorithm.  The reason why we are
rewriting it is that there are cases when the test is too short, and
we have observed that the result is more reliable if we insist for
more time.

Ideally, given an estimate of the available bandwidth and round-trip
time, we would like to fill the pipeline, possibly sending and receiving
many back-to-back small messages.  We have observed that this works
quite well with the under-development BitTorrent test.  Unluckily it
is not yet possible to do that with speedtest, because the current
http client code is based on the old stream, that does not have support
for request pipelining.  So, we will try to fill the pipeline using
the approach described here and implemented by this patch.

The general idea is that (1) you estimate the available bandwidth
first, with a quick test, and (2) then you choose a transfer size
that should keep the pipeline full for a reasonable time (we use
7 seconds).  We name `calibration` the first phase, and `measurement`
the second phase.

The calibration works as follows.  The first GET requests a body of
the minimum length, using range.  When we get the response, we adjust
the  length so that, given the elapsed time, if the measured speed is
correct, the download would span 1 second::

    self.length = int(self.length / elapsed)

then we calibrate again.  We stop calibrating if (i) the length
exceeds the maximum length, or (ii) the elapsed time is more than
a second, or (iii) we have already attempted to calibrate three
times.

When calibration is over, we adjust again the length so that, if
the measured speed is correct, the download with M connections would
span N seconds, e.g.::

    self.length = int(7 * self.length / 4)

in this case (the one of the download), we use 4 connections and we
want the test to span 7 seconds.  In the case of the upload, we instead
use just 2 connections, because, in the common case, the neubot agent
will be using an ADSL connection with poor upload speed.

The measurement is then the time required to transfer M times
self.length bytes.

The problem of this approach is that calibration might not work as
expected, for example because the connection does not exit slow-start,
and the consequence is that the test is shorter, and possibly slower,
than expected.  However, given the previous algorithm, that used a
fixed time, one second, per attempt, this one is certainly better,
because, on the average, the sustained flow will last more than one
second.