speedtest.py: Use 4 instead of 2 connections
This patch raises to four (from two) the number of connections we
use by default, to make the test more robust in case of lossy and/or
LFN paths.

This is the first patch in a branch that amends the speedtest algorithm
to increase robustness and reliability when the distance between neubot
agent and the server is high.  The branch consists of the following
patches:

1. speedtest.py: Use 4 instead of 2 connections
2. speedtest.py: Fist calibrate, then measure
3. speedtest.py: Measure download/upload speed every second
4. speedtest.py: upload: use 4 connections if needed

it also contains the following two small patches:

5. speedtest.py: Default to unit_formatter
6. speedtest.py: Use blupixel.net for on-demand tests

This patch just raises the number of connections but subsequent
patches will refine this a bit.  For example, if the upload path
is too slow, the current scheme, using two connections, would
be better.