www/results.html: Sync the about with index.html
While on that, let's say that, yes, it is painful to have to sync
the web pages because they share significant portions of their tag
soup.  But, luckily, soon there will be minimal support for SSI
(Server Side Includes) in-tree.  This will allow to factor out the
shared pieces, such as the header, the footer and the about.