bin/neubotw: Make sure stderr.write() is a no-op
When running under Win32, in windows mode, as a frozen executable:

a) we don't pass a DLL to NTEventLogHandler() with the proper magics
   that would allow to make slimmer logs, so the result is a bigger log
   file,

b) when the log file is full, windows would not write further logs,
   because the default (curious!) policy is that it would not overwrite
   logs older than 7 days,

c) so, the pywin32 code raises an exception, that cause handleError()
   to be invoked,

d) handleError() invokes traceback.print_exception() that prints on
   the standard error,

e) but the standard error, when running as a py2exe windows executable,
   is actually the file 'c:\program files\neubot\neubotw.exe.log',

f) if the log file was ever used, when the executable exists, it spawns
   an annoying popup suggesting the user to check out the logfile.

This branch aims to work-around the above issues in the quickest and
simplest way.  It is composed of two patches:

1. bin/neubotw: Make sure stderr.write() is a no-op
2. win32.py: Make sure we don't fill win32 logger

This patch applies a quick and dirty band-aid overriding the standard
error to make sure that stderr.write() is a no-op.  This should suffice
to avoid f).

The next patch suppresses info() and debug() when running as a windows
executable, in order to avoid b).