speedtest.py: Default to unit_formatter
By default use unit_formatter because this is the same formatter we
use in the Web User Interface (WUI), therefore one that reads the logs
can compare and match logs and WUI results.

While at that, also remember to add measurement unit to connect and
latency -- i.e. seconds -- because this was missing.