notify.py: Allow to pass publish() the opaque time
The following explanation might look a bit strange because it refers
to a set of enhancements that are still out-of-tree when this patch
is published.  You know... I am a modern Patch Processing Unit (PPU)
and I implement  `out of order commit`.

Anyway, here's the original explanation::

  We need to pass publish() the opaque time because publish() will
  cause STATE.marshal() to be invoked and so we must set the various
  `t` fields in state.py _before_.  And, in state.py, we MUST use
  the SAME opaque time we will then use in notify.py.  If we don't
  the comet mechanism will break because the time bound to the event
  in notify.py would not be same used in state.py -- in particular
  the latter will never be greater than the former.

An alternative explanation is that -- generally speaking -- it is
useful to know the time of an event before firing that event, expecially
if the time event is required by the event handling function code.