neubot/options.py: Remove __all__[]
The rationale of this change is that I want to avoid other modules
to import * from options (and to other modules) because importing everything
is implicit and it's better to do it explicit.  So, to discourage to
import everything from a module better not to provide an __all__[] list
at all.