neubot.js: Revert quick and dirty fix-update-box attempts
This commit reverts the following three commits (the most recent
one comes first):

a659891e... js/neubot.js: Make sure updateUrl is a link
7fc01f5f... js/neubot.js: Make sure updates box is visible
3b37d476... js/neubot.js: Make sure we output update URI

I rolled those commits out because I was on pressure for the release
and wanted to deploy a quick and dirty fix for the updates box.

Now, I need to back out because, in order to merge Alessio's patches,
I need to undo all the changes I applied while he was working.

(By the way, I bet my quick and dirty fix was not the right way to
 unbreak the update-box too).