net/stream.py: Become daemon on -Ddaemonize
This is the first patch in a branch that modifies net/stream.py and
bittorrent/main.py to be able to run in background and drop privileges,
if requested to do so.  This allows us to deploy experimental tests
for both TCP bulk transfer rate and BitTorrent on blupixel.

Patch details:

1. net/stream.py: Become daemon on -Ddaemonize
2. bittorrent/main.py: Become daemon on -Ddaemonize
3. main.py: Allow to invoke net/stream.py main()
4. bittorrent/main.py: Adjust --help message
5. net/stream.py: Adjust --help message
6. net/stream.py: Use 'stream' not 'net' prefix for options