speedtest.js: Very simple X axis labels rendering
We can do much better in this dept. and I recall Alessio having
a better diff than this one.  But, his diff has some issues that
need to be worked out, so I hope you don't protest too much if
I push this one waiting for the better one to be ready.