bittorrent/main.py: Delegate buffer scaling to the sender only
This commit is quite dirty and forces the receive buffer to be 262144
when we are the client, i.e. the one that downloads.  This basically
means that we delegate automatic buffer scaling to the sending operating
system only.  The rationale of this change is quite obscure.  I found
that this makes the test work better for me, but I need some more time
to figure out why.  So I am marking the relevant portion of code with
a big XXX because I may be wrong on that.  We'll see.