neubot.js: Ignore negotiate.queue_len
There is already a patch in (d421d70...) that amends index.html
not to show the number of Neubots in queue.  While reading neubot.js
before release, it makes sense of ignore negotiate.queue_len as
well.  As mentioned in d421d70..., this field is meaningless and
so better to dispose of it.