shell.py: Very simple command line shell
This merges some features of neubot/main.py and some features of
neubot/ui.py but with less overhead.  In particular, it doesn't
load unneeded modules, making the startup faster.  The support for
running in interactive mode is commented out because the rest of
the project is not ready for that.  Yes, that would be nice, but
there are tons of more important things to be done first.