boot.py: win32: Do not spawn a thread when already running
When neubot is already running, if we spawn a daemon thread, the
main thread of execution exits before the daemon one has had a chance
to start the web browser.  So, use the thread when we need to start
neubot (when we know the main thread of execution is not going to
exit soon) and just invoke the lambda when the daemon is already
running in background (indeed we don't need the create a new thread
in that case).