agent.py: The new default command
This command starts the web user interface, the local database, and
a rendezvous client.  This is same-as the current main() in rendezvous.py
but the difference is that agent.py is cleaner and focus just on the task
of being the agent, while rendezvous.py also contains the actual code
that deals with the rendezvous.