marshal.py: Allow to serialize/unserialize vectors too
This one is ugly.  It reminds me once again why JSON is better
than XML for many purposes.

But Neubot already deploys tons of ad-hoc XML marshallers and
unmarshallers.  And, in many cases, the capability to serialize
from/to vectors is required.

So, better to take the pill.  At least that allows to remove
piles of XML parsing code from rendezvous and speedtest.

Just a couple of additional notes:

1. serializing from/to JSON is the KISSest thing, so I'm deploying
   JSON unless that breaks backward compatibility;

2. my decision to switch from JSON to XML something like six months
   ago was _very_ unfortunate;

3. this commit removes the code that attempted to convert back
   to the original types, for the following reasons:

   3.1. strictly speaking this was not needed by our running code;

   3.2. that bit of code plus this one with vectors, and the routine
        for unmarshalling started looking too crappish;

   3.3. the absence of auto-casting should serve as a reminder that
        JSON is the long-term way to go.