rendezvous.py: Use marshal.py for RendezvousResponse
This time it's difficult to serialize and unserialize the data
structure using marshal.py XML capabilities.  But we cannot switch
abruptly to JSON for backward compatibility with already deployed
Neubots.

So, the server will send JSON to new clients, and XML to old clients
(i.e. < 0.3.7).  This way it suffices to write an ad-hoc marshaller
for the data structure: new clients will use JSON and old clients are
already deployed.

While here, this commit removes the code that checks for an alternate
server.  That code was an ad-hoc experiment carried out in December,
but we need much more work to support multiple servers in a clean way.
So better to clean the blackboard and restart from just one server.