database.py: Work-around timestamp being a float
In commit 7165c17... the following code snippet was removed from
the function that unmarshals from XML::

  -
  -        #XXX Workaround a bug in speedtest.py
  -        if element.tagName == "timestamp":
  -            cast = lambda x: int(float(x))
  -

The purpose of the snippet was to work-around a speedtest.py bug
that causes SpeedtestCollect timestamp to be a floating point instead
of an integer.  This was the right thing to do, but it was done in
the wrong place.  So, in the above-mentioned commit, I removed this
snippet from our generic marshaller.

But, the code in database.py assumes that ``timestamp`` is the string
serialization of an integer.  And, this is not true anymore because
we removed the check.  So, there is a TypeError exception and now we see
no results in speedtest.html because of that: the response to /api/speedtest
is now '500 Internal Server Error'.

The fix is as simple as converting the string to float() before trying
to coerce the timestamp to an integer, effectively adding the above check
again, but in another -- more proper -- place.