speedtest.py: Re-implement server-side using ServerHTTP
Note that the code that implements the real test is not a mixin of
the server anymore.  Instead now it follows the new ServiceHTTP pattern
supported by <neubot/http/server.py>.  That allows -- at least in
principle -- to host the test service into http/server.py for testing
purpose.  (Actually some more work is needed because the new service
still requires you to pass it a config object.)

Another notable hunk is the following one::

 @@ -271,8 +287,7 @@ class _NegotiateServerMixin(object):

      def _do_renegotiate(self, event, atuple):
          connection, request = atuple
 -        if connection.handler:
 -            self.do_negotiate(connection, request, True)
 +        self.do_negotiate(connection, request, True)

      def _speedtest_check_timeout(self):
          POLLER.sched(3, self._speedtest_check_timeout)

this should be safe now because publish() now wraps with try..except
the function it calls.  So we don't need to "ask for permission" before
invoking do_negotiate() anymore.