doc/speedtest-server.png: Dispose of now-obsolete documentation
Remove this piece of documentation because it's obsolete after
the recent changes.  Better to say nothing than to provide false
information to the interested reader.

I have the dia sources of that image off-tree.  When I will have
time I will update them to match the new organization of speedtest.py
and publish the image again on the repository.