net/stream.py: Simplify stream construction
This patch modifies the way we construct streams and propagate
changes to other files in the tree.  This lowers the overall
complexity, particularly in http/client.py and http/server.py.
The general idea of this patch is to avoid redundancy and to
avoid delegating tasks to lower level classes.