Elaborate on the idea of http server and child server
This patch extends the http/server.py model so that, given an http
server, it is possible to take another http server and redirect all
the requests beginning to a given prefix to this second server.

Before this patch this was possible for got_request() only while now
this is possible for got_request_headers() too.  This is very very
important for us, because it allows to host the whole speedtest server
as a child of another server.

The result is that now it is simpler to launch and test a server
from http/server.py.

Another notable change performed by this patch is that now we use
the configuration object that ships with every stream handler, hence
to each server, to configure the behavior.  This is important for
at least two reasons.

First we will use a key-value store bound to the configuration
database, in the future, instead of a plain dict and this will make
it possible to adjust the configuration on the fly.

Second because, leveraging the get() method of dicts, we are now
able to always provide defaults.  This is very valuable when we load
a server from http/server.py, without passing for the module object
and setting up the configuration properly.

The last change performed by this patch is that it rolls out the
much awaited new speedtest server, that should allow measurements
to take 'exactly N seconds'.  This is very important.

All of this is implemented in the same patch because I noticed that
the result is readable, so it seems ok to roll out the set of changes
instead of splitting the work in a serie of patches, i.e. there is
still a good signal/noise ratio.