speedtest.py: new complete speedtest client
This new complete speedtest client is based on the new http
client code and is significantly smaller and simpler than the
previous one.

It is possible to run this client via http/client.py in order
to test its behavior.  To do that, run something like this:

    $ python neubot/http.py -D class=ClientSpeedtest \
        -D module=neubot.speedtest http://your-uri/speedtest

Some more work needs to be done.  First we need to modify the
existing code to invoke this new client instead of the old one.
Then we need to add support for taking speed snapshots during
the transfers.

I guess the old client will go as soon as we're able to use
the new client.