speedtest.py: start using new speedtest client
This patch is very quick and dirty because it creates yet another
wrapper of the SpeedtestClient and adds lines to a file that already
count too many lines.  But the faster this goes out the better.

I will schedule some time to cleanup the content of speedtest.py
and to garbage collect other old files.