speedtest.py: don't connect() all connections at the same time
If we connect() all connections at the same time, the RTT estimate
is less accurate.  For certain connections, indeed, we also account
for the time spent waiting for other connections to resolve the
domain name and connect.

Now the resulting RTT should be more reliable since we connect() one
connection at a time.