My dying neubot/pathnames.py
The way the configuration works is to be changed because the conf
should go into a table in the database (Alessio Palmero, you finally
managed to change my mind on that!)

This means we need to rewrite the configuration bits of all mains
of all modules.

And pathnames.py happens to be used only by the parts that we need
to rewrite.

So, to accellerate pathnames.py death, we stop using it.