log.py: Rewrite that crappy serialization mechanism
The previous serialization mechanism was crappish for at least two
reasons.  First, it does not make sense for str(LOG) to return a JSON
object, because str() is supposed to return a *string*.  Second, it
is not necessary to build a dictionary per log entry, since a list
would be enough for this use case.