Do not change directory to profile dir
Certainly it is saner to use an absolute path for the database instead
of (i) chdir() to profile dir and then (ii) use a relative path.

So add system/ functions to (a) return the default database file name,
and (b) ensure the database file path is absolute, exists, and has assigned
the right permissions.

Change_dir() is still useful, tho.  Just modify its semantic to switch
to the root directory (under UNIX only).

Also, modify agent.py to change_dir() when running in background only.

Also, ensure that database/__init__.py calls (a) and (b).