utils.py: Make sure we use integers to represent booleans
One of the ideas behind config.py is *type persistence*.  The general
idea of that is the following.  You set the value of a property for
the first time.  And this defines the property type forever.  If,
later on, you reassign such property to another value, config.py do
guarantee that the new value has the same type of the old value.

The current implementation of this idea has the problem that we
cannot represent booleans as strings.  So, this patch enhances the
code in utils.py to map common string representation of booleans
to integer booleans.  E.g., now "on" maps to 1.