Keep boolean values consistent across CONFIG lifetime
The API uses integers to represent booleans.  In particular, zero
is false, and non-zero is true.  As usual.

For additional robustness, utils.py contains code that converts boolean
strings directly to integers.  So the following snippet will work::

    $ cat foo.py

    ...
    CONFIG.register_defaults({
        "foo.bar": 1,
    })
    ...

    $ neubot foo -v -Dbar=off
    ...
    config: foo.bar: 1 -> 0

This works because "off" becomes 0 in utils.py:intify().

We now want to guarantee that GET /api/config yields the same
"foo.bar" value, whether the property has assigned the default
or not.  This is needed to avoid inconsistency::

    foo.bar="True"    # using the default value
    foo.bar=0         # when modified via -D

This is likely to cause headaches in the javascript dept.  To avoid
that, just apply the cast the first time we set the property value.
In most cases that would be a no-op, but in the boolean case it will
convert booleans to integers::

    CONFIG.register_defaults({
        "foo.n": 3,           # same as before
        "foo.s": "abc",       # same as before
        "foo.bar": True,      # True -> 1
    })

In the end, this allows the neubot python programmer to deploy
some more syntactic sugar.  Now it is possible to default properties
to True or False, but the code will automagically convert that to
integers.