Don't store reference to measurer into configuration
The code in config.py and utils.py that guarantees that properties
keep the same type during time is neither able to handle None nor able
to handle a reference to a measurer object.

So modify net/stream.py API to avoid storing a reference to the
current measurer into the configuration.