speedtest: force the measurer we want just before we need it
This is the proper fix to the measurer problem, resulting in a
stacktrace when running the speedtest.

The problem is that commit bf0299a6... modifies the way we
pass the measurer down the chain.  In particular attach() now
forces the measurer.

So, the measurer set by the initializer of ClientSpeedtest
is lost during attach().

A previous patch already proposed a quick fix for the problem
but this one is more correct because it moves the construction
of the measurer we want were it's not possible for the code
in HTTP and stream to mistakenly override it.

While there, also be correct and set the measurer=None argument
in configure() in speedtest/__init__.py