notify.py: TESTDONE should not run periodically
So far we have used notify.py mainly for the comet and for
renegotiate.  For both events, after a given period it make
sense to fire the event nonetheless.  To keep respectively
the browser and the clients in queue informed that we are
still alive, even if nothing has happened.

But TESTDONE is a different beast.  In this case it's not
such a good idea to fire it periodically.  We must wait for
the test to actually stop.

This requires to improve the design of notify.py.  But not
now.  For now, it's OK to special-case it.