net/stream.py: separate handling of send and recv socket buffers
We force the receive buffer to 262144 to be sure that the receiving
operating system is not going to inflate it too slowly.  This will be
a problem because we want the test to be quick, so we don't want to
loose time because of receiver-buffer-limited slow-start.

For the send buffer we let the sending operating system scale the
size.  The smooth scaling should guarantees that the sender does not
open its send window too quickly.

(The receiver does not have information on the actual congestion
window, while the sender of course has that.  For this reason the
scaling is simpler on the sender side.)