net/stream.py: allow to tune and reduce default chunk size
Allow to tune the chunk size.  We define chunk the amount of bytes
that is written by the chargen test per time.

Reduce the default chunk size.  Now it is close to the one used by
BitTorrent.  The previous default was good for loopback testing, but
I suspect the new value would be enough for real networks.  If I'm
wrong, I would bump it up again, and, meanwhile, there is the option
to adjust it.