Move neubot/options.py into the attic
The last user of neubot/options.py has been switched to config.py
so we can now decommission the former.

BTW neubot/options.py has been very important for Neubot because it
was the first successful attempt at tackling the complexity of a growing
set of options.

Also move to the attic the manual page doc/neubot.3.txt because it
contains only documentation of options.py and that is clearly becoming
very obsolete.