config.py: do not return ConfigDict() on select()
It's not feasible to return ConfigDict() on select() because the
user of such dictionary does not want to restrict to simple types
for this dictionary.  For example, the speedtest test wants to use
this dictionary to store intermediate results and save on variables
(and also on complexity).  This is a good thing, and I think we
should allow it.

The drawback is that the type-safety does not propagate to dicts
returned by select().  I'm not sure that this is an issue: after
all, the place where type-safety is more needed is where there is
interaction with users.

This problem was not visibile before because http/client.py has
been based on options.py until yesterday.  So the dictionary was
indeed just a dictionary.  Today I've started testing the code
in speedtest/client.py with http/client.py using config.py.  And
so I noticed the problem.