log.py: don't flood log with access logs
The web user interface now exports the logs.  This is very cool
but there is a problem.  This page is flooded by access logs generated
via comet by the user interface itself.  A quick and dirty solution
to this problem is the one provided by this patch.  Do not export via
API access logs.