api/server.py: Make /api more transparent and debuggable
This patch makes /api more transparent and debuggable.  We allow
all APIs to receive an optional parameter debug=BOOL.  Whenever
this parameter is true, we return an indented JSON string with mime-
type 'text/plain'.  So you can inspect the state directly from the
browser.

Note that /api/log?debug=1 does not return a JSON but rather
plain text.  This is because it's easier to read line-based logs
rather than eye-parsing the formatted JSON.

While there, this patch also harmonizes all the modules that are
invoked by the API server.  In particular, now, they all return
objects.  It is API server responsibility to marshal them according
to the user-specified policy.

I need this patch right now because I have to make Neubot more
transparent in order to understand why it blocked under Windows
in rendezvous state.