api/server.py: return available APIs on GET /api
This patch modifies API server to return a list of available APIs
if one GETs /api.  This is reasonable because allows one to discover
and list all the available APIs.  For example, sometimes I don't
recall the names and it's a pain to less neubot/api/server.py just
to do that.