net: s/closing/closed/ in Connector and Listener
From time to time the release candidate got stuck in rendezvous
state and stayed there forever.  Thanks to new /api/debug facilities
I was able to see that: (i) RendezvousClient is waiting forever for
'testdone'; and (ii) the last state update is about the test name.
So the problem should be when connecting to the remote server.

Looking at the source code, I noticed the bug.  The poller invokes
the closed() function in case of timeout.  But the Connector does not
define such function and, instead, defines closing().  So, instead,
the default no-op Pollable.closed() method is invoked and indeed nothing
actually happens.  The fix is s/closing/closed/ in Connector.

While there, also fix Listener, that has the same bug.