api/server.py: import version from boot
Neubot/boot.py is the common main() of most Neubot commands, and
it keeps the current version number.  We can pull the version number
from that.  This has no loading-yet-another-module penalities because
neubot/boot.py is already loaded in any case (unless you python(1)
neubot/api/server.py but that is not the common use case).