rendezvous: move message-handling code to compat.py
This commit moves message-handling code to compat.py without making
any functional change.  The underlying idea is to separate the handling
of messages from the rest of the rendezvous code.  In particular, in
my opinion it would be nice to use dictionaries in rendezvous and: (i)
use JSON with new clients; (ii) use compat.py to transform dictionaries
to XML with obsolete ones.