rendezvous/server.py: use releases.n.o instead of w.n.o/download
The rationale of this change is that using releases.n.o the automatic
downloader can construct easily the URI to download from, given the
target version number.  This is much more difficult if instead the base
URI is w.n.o/download because in this case the downloader need to read
and parse the returned HTML page.