http/client.py: make connect_uri() more robust
In particular, generate connection_failed() if connect_uri() execepts
for nearly any reason.  The exception will then be handled upstream.

This is better than not catching the exception, causing the loss of
the related socket and possibly of one event too.

If the URI is wrong, loosing the socket and the event might cause Neubot
to stay forever in the rendezvous state.  Not pleasant.