rendezvous/client.py: subscribe before connect_uri
So, if connect_uri() fails we're in a good shape, we see
immediately the failure and we can schedule another attempt.
Before this patch, the error arrived _before_ we had a
chance of registering for TESTDONE event.  So, connection
failed, the event was fired, no one cared, and, eventually,
after that we registered for that event.

Without this patch Neubot can stay forever in the rendezvous
state.  Sigh.