api/server.py: Start moving properties to config.py
It's paramount that all the properties have their default value
before we merge values from the outside world, or strange things
are likely to happen.

This patch moves privacy properties from server.py to config.py.
We start with privacy properties because they are one of the most
important settings here in Neubot.  So we want to be sure there
are no surprises.

Spotted running a bare speedtest client against a just modified
speedtest negotiator.  Woah!  To make many changes is really the
patch to find a lot of bugs everywhere.