speedtest: cast privacy settings to integers
So we are sure that they are evaluated as expected.  Here instead
the problem is that the server receives strings due to the in-the-middle
XML layer.  Once again, JSON would have been simpler.  However, the
cast to integer fixes the issue for me.  The client has not such issues
but for simmetry I'm deploying the same countermeasures there too.
Cut-n-paste is always possible, expecially many months later.